requirements.txt
asgiref==3.2.7
bcrypt==3.1.7
bson==0.5.8
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
dj-database-url==0.5.0
Django==3.0.5
django-forms-bootstrap==3.1.0
django-mathfilters==1.0.0
django-taggit==1.2.0
dnspython==1.16.0
heroku==0.1.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
pycparser==2.19
pymongo==3.9.0
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.1
urllib3==1.25.6
Werkzeug==0.16.0
WTForms==2.2.1
